+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning].
 ## Unreleased
 
 ### Removed
-- `iam tenant` commands have been removed.
+- `iam tenant` commands have been removed. Use `iam privilege` instead.
 
 ### Changed
 - `iam privilege` operates on global privileges by default; `--record-type` and `--global-id` can be specified otherwise.
@@ -21,6 +21,11 @@ and this project adheres to [Semantic Versioning].
 - `iam privilege list-by-user` now accepts `-gid, --global-id` rather than a positonal argument.
 - `org privilege list-by-role` now accepts `-gid, --global-id` rather than a positonal argument.
 - `org privilege list-by-user` now accepts `-gid, --global-id` rather than a positonal argument.
+## [0.75.0] - 2025-11-05
+
+### Changed
+- `sample [share|unshare]` can now take the project abbrev instead of group name as shorthand.
+- `sample [chown|share|unshare|disable|enable]` can now either accept multiple `--seq-id` flags, or `--file` with a file containing a list of samples.
 
 ## [0.74.0] - 2025-10-19